Patch for qprintable crash on Debian wheezy
On Debian wheezy the executable "karlyriceditor" crashes on startup with the following message:
> QThread: Destroyed while thread is still running
> Segmentation fault
The reason was an inadequate use of the (dangerous) qPrintable macro. The macro evaluates to a pointer that points to memory within a temporary object. So this pointer becomes invalid as soon as the whole expression is evaluated. If the pointer has to be used outside of this expression then either its content has to be copied (and deleted when no longer needed) or (what is done in this patch) the temporary object has to be stored in a local variable.