Patch to fix compilation: Translate FFmpeg calls to Libav-compatible calls
The upstream developers compile Karlyriceditor against FFmpeg, but in Debian there is only Libav available, whose API is slowly drifting away from FFmpeg's. This patch adapts the code to 2 of those differences.
Firstly the second parameters of "avformat_new_stream" is of non-const pointer type, but the code tries to pass a const pointer to it.
Secondly there is no AV_ROUND_PASS_MINMAX flag for the rounding parameter of av_rescale_q_rnd in Libav.