Fix floating-point exception due to unset sample aspect ratio of AVStream (has to be the same as the one of the codec in Libav, upstream uses FFmpeg)
The upstream developers compile Karlyriceditor against FFmpeg, but in Debian there is only Libav available, whose API is slowly drifting away from FFmpeg's. This patch adapts the code to one of those differences. In Libav sample aspect ratio of a AVStream object is not automatically initialized to the one of its codec. So this has to be done manually.